Rename mesh_sampler to mesh_ops.
